Adding media id to like endpoint.
Without the id we would not be able to differentiate between different media with the same or similar names.
But id may be the same across media types, so we need the name and id to be sure.

(pod with name "name" and id "8" could otherwise be echanged with video "other name" with id "8")